encapsulate prosemirror
